+++ src/Model.php
@@ -602,7 +602,7 @@ abstract class Model implements \ArrayAccess
             return $this->create();
         }
 
-        return $this->set($this->_unsaved);
+        return $this->set();
     }
 
     /**
@@ -620,11 +620,8 @@ abstract class Model implements \ArrayAccess
             throw new BadMethodCallException('Cannot call create() on an existing model');
         }
 
-        if (!empty($data)) {
-            foreach ($data as $k => $value) {
-                $this->$k = $value;
-            }
-        }
+        // mass assign values passed into create()
+        $this->setValues($data);
 
         // dispatch the model.creating event
         if (!$this->handleDispatch(ModelEvent::CREATING)) {
@@ -795,6 +792,23 @@ abstract class Model implements \ArrayAccess
     }
 
     /**
+     * Sets a collection values on the model from an untrusted input.
+     *
+     * @param array $values
+     *
+     * @return self
+     */
+    public function setValues($values)
+    {
+        foreach ($values as $k => $value) {
+            // TODO check for mass assignment violations
+            $this->$k = $value;
+        }
+
+        return $this;
+    }
+
+    /**
      * Converts the model to an array.
      *
      * @return array
@@ -838,16 +852,12 @@ abstract class Model implements \ArrayAccess
             throw new BadMethodCallException('Can only call set() on an existing model');
         }
 
-        // not updating anything?
-        if (count($data) == 0) {
-            return true;
-        }
+        // mass assign values passed into set()
+        $this->setValues($data);
 
-        // apply mutators
-        foreach ($data as $k => $value) {
-            if ($mutator = self::getMutator($k)) {
-                $data[$k] = $this->$mutator($value);
-            }
+        // not updating anything?
+        if (count($this->_unsaved) == 0) {
+            return true;
         }
 
         // dispatch the model.updating event
@@ -856,14 +866,14 @@ abstract class Model implements \ArrayAccess
         }
 
         // DEPRECATED
-        if (method_exists($this, 'preSetHook') && !$this->preSetHook($data)) {
+        if (method_exists($this, 'preSetHook') && !$this->preSetHook($this->_unsaved)) {
             return false;
         }
 
         // validate the values being saved
         $validated = true;
         $updateArray = [];
-        foreach ($data as $name => $value) {
+        foreach ($this->_unsaved as $name => $value) {
             // exclude if value does not map to a property
             if (!isset(static::$properties[$name])) {
                 continue;